// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
	ApiFromModules,
	FilterApi,
	FunctionReference,
} from "convex/server";
import type * as habits from "../habits.js";
import type * as habitLogs from "../habitLogs.js";
import type * as users from "../users.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
	habits: typeof habits;
	habitLogs: typeof habitLogs;
	users: typeof users;
}>;
export declare const api: FilterApi<
	typeof fullApi,
	FunctionReference<"action" | "mutation" | "query", "public">
>;
export declare const internal: FilterApi<
	typeof fullApi,
	FunctionReference<"action" | "mutation" | "query", "internal">
>;